README.md
# tiddler-expedition
# introduction
this is a demo game
# rules of the game
First of all, you have to enter the number of rows and columns in the game box. The number of rows and columns is limited to 6 to 15.
Then, a game box can be generated. At first, you are a small fish in the upper left corner. Click the Start Game button to get the initial time of 66 seconds.
If you eat shrimp, you can get 8 to 11 points; If you get the treasure chest, you can get 10 points and add 6 seconds. If an anchor point is encountered, the game ends.
Go and experience it! !